Moving submission create form back to page
